Create LoginPage.sol
Users can submit their fingerprints and facial recognition data via the user interface.
They can take a picture of their face and a picture of their fingerprints, or upload images of their fingerprints and facial recognition data.
Once submitted, the data is sent to the smart contract's requestAccess function, where it is verified by calling the checkAccess function. If the user is verified, the AccessGranted event is emitted, otherwise the AccessDenied event is emitted.